+++ apps/accounts/forms/__init__.py
@@ -6,6 +6,7 @@ from .login import LoginForm
 from .password_reset import CaptchaPasswordResetForm
 from .profile import AvatarForm, ProfileForm
 from .register import DynamicStepForm, EmailStepForm, PasswordStepForm
+from .totp import TOTPDisableForm, TOTPSetupForm, TOTPVerifyForm
 
 __all__ = [
     "AvatarForm",
@@ -15,4 +16,7 @@ __all__ = [
     "LoginForm",
     "PasswordStepForm",
     "ProfileForm",
+    "TOTPDisableForm",
+    "TOTPSetupForm",
+    "TOTPVerifyForm",
 ]